# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AiSummaryOut": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "const": "NOT_ENABLED",
            "default": "NOT_ENABLED",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AiSummaryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HandoffOut": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instructions": {
            "title": "Instructions",
            "type": "string"
          },
          "next_step_code": {
            "const": "CONFIRM_HANDOFF",
            "title": "Next Step Code",
            "type": "string"
          },
          "next_step_title": {
            "title": "Next Step Title",
            "type": "string"
          },
          "owner_display_name": {
            "title": "Owner Display Name",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "status": {
            "const": "READY",
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "owner_user_id",
          "owner_display_name",
          "title",
          "next_step_code",
          "next_step_title",
          "instructions",
          "created_at"
        ],
        "title": "HandoffOut",
        "type": "object"
      },
...
      "NotificationDeliveryOut": {
        "additionalProperties": false,
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "delivery_scope": {
            "const": "LOCAL_TEST_ONLY",
            "default": "LOCAL_TEST_ONLY",
            "title": "Delivery Scope",
            "type": "string"
          },
          "display_status": {
            "title": "Display Status",
            "type": "string"
          },
          "external_delivery_confirmed": {
            "const": false,
            "default": false,
            "title": "External Delivery Confirmed",
            "type": "boolean"
          },
          "last_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Code"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "channel",
          "display_status",
          "attempt_count",
          "next_attempt_at",
          "last_error_code",
          "delivered_at"
        ],
        "title": "NotificationDeliveryOut",
        "type": "object"
      },
...
      "ResultEvaluationOut": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "const": "PASS",
            "title": "Decision",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "overall_feedback": {
            "title": "Overall Feedback",
            "type": "string"
          },
          "reviewer_id": {
            "format": "uuid",
            "title": "Reviewer Id",
            "type": "string"
          },
          "rubric_feedback": {
            "items": {
              "$ref": "#/components/schemas/ResultRubricFeedbackOut"
            },
            "title": "Rubric Feedback",
            "type": "array"
          }
        },
        "required": [
          "id",
          "reviewer_id",
          "decision",
          "overall_feedback",
          "rubric_feedback",
          "created_at"
        ],
        "title": "ResultEvaluationOut",
        "type": "object"
      },
      "ResultOut": {
        "additionalProperties": false,
        "properties": {
          "ai_summary": {
            "$ref": "#/components/schemas/AiSummaryOut"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision": {
            "const": "PASS",
            "title": "Decision",
            "type": "string"
          },
          "evaluation": {
            "$ref": "#/components/schemas/ResultEvaluationOut"
          },
          "handoff": {
            "$ref": "#/components/schemas/HandoffOut"
          },
          "notification": {
            "$ref": "#/components/schemas/NotificationDeliveryOut"
          },
          "outcome_id": {
            "format": "uuid",
            "title": "Outcome Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "outcome_id",
          "decision",
          "status",
          "summary",
          "evaluation",
          "handoff",
          "notification",
          "ai_summary",
          "created_at"
        ],
        "title": "ResultOut",
        "type": "object"
      },
      "ResultResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ResultOut"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "ResultResponse",
        "type": "object"
      },
      "ResultRubricFeedbackOut": {
        "additionalProperties": false,
        "properties": {
          "dimension_key": {
            "title": "Dimension Key",
            "type": "string"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "rating": {
            "title": "Rating",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "dimension_key",
          "title",
          "rating",
          "feedback"
        ],
        "title": "ResultRubricFeedbackOut",
        "type": "object"
      },
...
      "TimelineItemOut": {
        "additionalProperties": false,
        "properties": {
          "details": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Details",
            "type": "object"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "object_id": {
            "format": "uuid",
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "title": "Object Type",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "event_type",
          "title",
          "occurred_at",
          "object_type",
          "object_id",
          "details"
        ],
        "title": "TimelineItemOut",
        "type": "object"
      },
      "TimelineOut": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "TimelineOut",
        "type": "object"
      },
      "TimelineResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TimelineOut"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "TimelineResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Muchen Journey vNext API",
    "version": "0.2.0-local"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/result": {
      "get": {
        "operationId": "result_api_v1_me_result_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fixture-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fixture-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Result"
      }
    },
...
    "/api/v1/me/timeline": {
      "get": {
        "operationId": "timeline_api_v1_me_timeline_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 500,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Fixture-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fixture-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timeline"
      }
    },
...
  }
}